+++ Order-Core/src/main/java/com/goormgb/be/ordercore/match/dto/MatchGuideDto.java
@@ -1,14 +1,14 @@
 package com.goormgb.be.ordercore.match.dto;
 
-import com.goormgb.be.ordercore.match.enums.PurchaseStatus;
+import com.goormgb.be.domain.match.enums.PurchaseStatus;
 
 public record MatchGuideDto(
-		String teamsDisplay,
-		String ageLimit,
-		String placeDisplay,
-		String addressDisplay,
-		String datetimeDisplay,
-		PurchaseStatus purchaseStatus,
-		String matchDdayLabel
+	String teamsDisplay,
+	String ageLimit,
+	String placeDisplay,
+	String addressDisplay,
+	String datetimeDisplay,
+	PurchaseStatus purchaseStatus,
+	String matchDdayLabel
 ) {
 }
