+++ Order-Core/src/main/java/com/goormgb/be/ordercore/match/service/MatchService.java
@@ -9,15 +9,17 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.goormgb.be.domain.club.entity.Club;
+import com.goormgb.be.domain.match.entity.Match;
+import com.goormgb.be.domain.match.repository.MatchRepository;
 import com.goormgb.be.global.exception.ErrorCode;
 import com.goormgb.be.global.support.Preconditions;
-import com.goormgb.be.ordercore.club.entity.Club;
+
 import com.goormgb.be.ordercore.club.repository.ClubRepository;
 import com.goormgb.be.ordercore.match.dto.response.ClubMonthlyMatchesResponse;
 import com.goormgb.be.ordercore.match.dto.response.MatchDetailGetResponse;
 import com.goormgb.be.ordercore.match.dto.response.MatchListByDateResponse;
-import com.goormgb.be.ordercore.match.entity.Match;
-import com.goormgb.be.ordercore.match.repository.MatchRepository;
+
 import com.goormgb.be.ordercore.match.utils.MatchDisplayUtils;
 import com.goormgb.be.ordercore.match.utils.SalesOpenUtils;
 
@@ -27,10 +29,10 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Transactional
 public class MatchService {
-	final private MatchRepository matchRepository;
-	final private ClubRepository clubRepository;
-	final private MatchDisplayUtils matchDisplayUtils;
-	final private SalesOpenUtils salesOpenUtils;
+	private final MatchRepository matchRepository;
+	private final ClubRepository clubRepository;
+	private final MatchDisplayUtils matchDisplayUtils;
+	private final SalesOpenUtils salesOpenUtils;
 
 	public MatchDetailGetResponse getMatchDetail(Long id) {
 		var match = matchRepository.findDetailByIdOrThrow(id);
@@ -44,11 +46,11 @@ public class MatchService {
 		Instant end = date.plusDays(1).atStartOfDay(ZoneOffset.UTC).toInstant();
 
 		List<Match> matches = matchRepository.findAllByMatchAtGreaterThanEqualAndMatchAtLessThanOrderByMatchAtAsc(start,
-				end);
+			end);
 
 		var summaries = matches.stream()
-				.map(m -> MatchListByDateResponse.MatchSummary.of(m, salesOpenUtils.calculateSalesOpenAt(m)))
-				.toList();
+			.map(m -> MatchListByDateResponse.MatchSummary.of(m, salesOpenUtils.calculateSalesOpenAt(m)))
+			.toList();
 
 		return MatchListByDateResponse.of(date, summaries);
 	}
@@ -65,8 +67,8 @@ public class MatchService {
 		List<Match> matches = matchRepository.findMonthlyByClubId(clubId, start, end);
 
 		List<ClubMonthlyMatchesResponse.MatchItem> items = matches.stream()
-				.map(m -> toItem(clubId, m))
-				.toList();
+			.map(m -> toItem(clubId, m))
+			.toList();
 
 		return ClubMonthlyMatchesResponse.of(clubId, year, month, items);
 	}
@@ -76,15 +78,15 @@ public class MatchService {
 		Club opponent = isHome ? m.getAwayClub() : m.getHomeClub();
 
 		return new ClubMonthlyMatchesResponse.MatchItem(
-				m.getId(),
-				m.getMatchAt(),
-				new ClubMonthlyMatchesResponse.OpponentClub(
-						opponent.getId(),
-						opponent.getKoName(),
-						opponent.getLogoImg()
-				),
-				m.getSaleStatus(),
-				isHome
+			m.getId(),
+			m.getMatchAt(),
+			new ClubMonthlyMatchesResponse.OpponentClub(
+				opponent.getId(),
+				opponent.getKoName(),
+				opponent.getLogoImg()
+			),
+			m.getSaleStatus(),
+			isHome
 		);
 	}
 }
